Update make_bulk_fastq.pl (#2)
For the first time, I ran BAM files containing only unmapped reads. In this case the “properly paired” field reports 0, which makes sense because that counter is based on the 0x2 flag, since all reads are unmapped (0x4 set). I know the reads are paired as the reads flag contain 0x1, and that shows up correctly under the “paired in sequencing” line in the flagstat output.